requirements_trading.txt
# 빗썸 자동매매 필수 라이브러리
pybithumb>=1.0.0          # 빗썸 API 래퍼
pandas>=1.5.0             # 데이터 처리
numpy>=1.23.0             # 수치 계산
schedule>=1.2.0           # 스케줄링
python-dotenv>=1.0.0      # 환경변수 관리
requests>=2.28.0          # HTTP 요청
